+++ js/game.js
@@ -1,10 +1,9 @@
 var player;
 var needCanvasUpdate = true;
-var gameEnded = false;
 
 // Don't change this
 const TMT_VERSION = {
-	tmtNum: "2.6.1",
+	tmtNum: "2.6.4.3",
 	tmtName: "Fixed Reality"
 }
 
@@ -114,8 +113,7 @@ function canReset(layer){
 function rowReset(row, layer) {
 	for (lr in ROW_LAYERS[row]){
 		if(layers[lr].doReset) {
-
-			Vue.set(player[lr], "activeChallenge", null) // Exit challenges on any row reset on an equal or higher row
+			if (!isNaN(row)) Vue.set(player[lr], "activeChallenge", null) // Exit challenges on any row reset on an equal or higher row
 			run(layers[lr].doReset, layers[lr], layer)
 		} else if (tmp[layer].row > tmp[lr].row && row !== "side" && !isNaN(row)) {
 			layerDataReset(lr)
@@ -130,19 +128,17 @@ function layerDataReset(layer, keep = []) {
 		if (player[layer][keep[thing]] !== undefined)
 			storedData[keep[thing]] = player[layer][keep[thing]]
 	}
+
 	Vue.set(player[layer], "buyables", getStartBuyables(layer))
 	Vue.set(player[layer], "clickables", getStartClickables(layer))
 	Vue.set(player[layer], "challenges", getStartChallenges(layer))
+	Vue.set(player[layer], "grid", getStartGrid(layer))
 
 	layOver(player[layer], getStartLayerData(layer))
 	player[layer].upgrades = []
 	player[layer].milestones = []
 	player[layer].achievements = []
-	player[layer].challenges = getStartChallenges(layer)
-	resetBuyables(layer)
 
-	if (layers[layer].clickables && !player[layer].clickables) 
-		player[layer].clickables = getStartClickables(layer)
 	for (thing in storedData) {
 		player[layer][thing] =storedData[thing]
 	}
@@ -164,24 +160,23 @@ function generatePoints(layer, diff) {
 	addPoints(layer, tmp[layer].resetGain.times(diff))
 }
 
-var prevOnReset
-
 function doReset(layer, force=false) {
 	if (tmp[layer].type == "none") return
 	let row = tmp[layer].row
 	if (!force) {
+		
+		if (tmp[layer].canReset === false) return;
+		
 		if (tmp[layer].baseAmount.lt(tmp[layer].requires)) return;
 		let gain = tmp[layer].resetGain
 		if (tmp[layer].type == "static") {
 			if (tmp[layer].baseAmount.lt(tmp[layer].nextAt)) return;
 			gain = (tmp[layer].canBuyMax ? gain : 1)
 		} 
-		if (tmp[layer].type == "custom") {
-			if (!tmp[layer].canReset) return;
-		} 
+
+		if (layers[layer].onPrestige) run(layers[layer].onPrestige, layers[layer], gain)
 		
 		addPoints(layer, gain)
-		if (layers[layer].onPrestige) run(layers[layer].onPrestige, layers[layer], gain)
 		updateMilestones(layer)
 		updateAchievements(layer)
 
@@ -196,21 +191,21 @@ function doReset(layer, force=false) {
 			}
 		}
 	
-		tmp[layer].baseAmount = decimalZero // quick fix
 	}
 
 	if (run(layers[layer].resetsNothing, layers[layer])) return
+	tmp[layer].baseAmount = decimalZero // quick fix
 
 	for (layerResetting in layers) {
 		if (row >= layers[layerResetting].row && (!force || layerResetting != layer)) completeChallenge(layerResetting)
 	}
 
-	prevOnReset = {...player} 
 	player.points = (row == 0 ? decimalZero : getStartPoints())
 
 	for (let x = row; x >= 0; x--) rowReset(x, layer)
-	rowReset("side", layer)
-	prevOnReset = undefined
+	for (r in OTHER_LAYERS){
+		rowReset(r, layer)
+	}
 
 	player[layer].resetTime = 0
 
@@ -302,15 +297,15 @@ function autobuyUpgrades(layer){
 }
 
 function gameLoop(diff) {
-	if (isEndgame() || gameEnded){
-		gameEnded = 1
+	if (isEndgame() || tmp.gameEnded){
+		tmp.gameEnded = true
 		clearParticles()
 	}
 
 	if (isNaN(diff) || diff < 0) diff = 0
-	if (gameEnded && !player.keepGoing) {
+	if (tmp.gameEnded && !player.keepGoing) {
 		diff = 0
-		//player.tab = "gameEnded"
+		//player.tab = "tmp.gameEnded"
 		clearParticles()
 	}
 
@@ -380,7 +375,7 @@ var pastTickTimes = []
 function runInterval(){
 	if (player === undefined || tmp === undefined) return;
 	if (ticking) return;
-	if (gameEnded && !player.keepGoing) return;
+	if (tmp.gameEnded && !player.keepGoing) return;
 	ticking = true
 	let now = Date.now()
 	let diff = (now - player.time) / 1e3